Move bot/vote_session.py and database/vote.py into database/voting
