+++ bot/submission/submit.py
@@ -22,7 +22,7 @@ import vecs
 
 from bot import utils
 from bot.submission.parse import parse_dimensions
-from bot.vote_session import AbstractVoteSession
+from database.voting.vote_session import AbstractVoteSession
 from bot.submission.ui import BuildSubmissionForm, ConfirmationView
 from database import message as msg
 from database.builds import get_all_builds, Build
@@ -32,7 +32,7 @@ from bot._types import GuildMessageable
 from bot.utils import RunningMessage, is_owner_server, check_is_staff, check_is_trusted_or_staff, is_staff
 from database.schema import TypeRecord, RestrictionRecord, RestrictionAliasRecord
 from database.utils import upload_to_catbox
-from database.vote import track_build_vote_session, track_vote_session, close_vote_session
+from database.voting.vote import track_build_vote_session, track_vote_session, close_vote_session
 
 if TYPE_CHECKING:
     from bot.main import RedstoneSquid
