+++ database/voting/vote.py
